requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeladmin-reorder==0.3.1
django-non-dark-admin==2.0.2
Pillow==9.4.0
sqlparse==0.4.3
